logistic regression demo added
